+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Dev: Update to Twitter's v2 API. (#414)
 - Dev: Add HTTP Caching headers. (#417)
 - Dev: Add custom middleware to strip trailing slashes. (#422)
+- Dev: Make cache timeout durations configurable. (#419)
 
 ## 1.2.3